+++ apps/kyb-app/src/domains/collection-flow/collection-flow.api.ts
@@ -1,19 +1,22 @@
+import get from 'lodash/get';
+import posthog from 'posthog-js';
+
 import { request } from '@/common/utils/request';
+import { CollectionFlowContext } from '@/domains/collection-flow/types/flow-context.types';
 import {
+  CreateEndUserDto,
   DocumentConfiguration,
+  EndUser,
+  FlowContextResponse,
   IDocumentRecord,
   TCustomer,
   TFlowConfiguration,
   TFlowStep,
   TUser,
   UISchema,
+  TUpdateEndUserPluginData,
+  TFetchCompanyInformationPluginData,
 } from '@/domains/collection-flow/types';
-import {
-  CollectionFlowConfig,
-  CollectionFlowContext,
-} from '@/domains/collection-flow/types/flow-context.types';
-import get from 'lodash/get';
-import posthog from 'posthog-js';
 
 export const fetchUser = async (): Promise<TUser> => {
   const user = await request.get('collection-flow/user').json<TUser>();
@@ -72,10 +75,13 @@ export const fetchCustomer = async (): Promise<TCustomer> => {
   return await request.get('collection-flow/customer').json<TCustomer>();
 };
 
-export interface FlowContextResponse {
-  context: CollectionFlowContext;
-  config: CollectionFlowConfig;
-}
+export const fetchWorkflowId = async (token: string | null): Promise<string> => {
+  if (!token) {
+    return '';
+  }
+
+  return await request.get(`collection-flow/workflow-id`).text();
+};
 
 export const fetchFlowContext = async (): Promise<FlowContextResponse> => {
   try {
@@ -93,25 +99,9 @@ export const fetchFlowContext = async (): Promise<FlowContextResponse> => {
   }
 };
 
-export interface EndUser {
-  id: string;
-  email: string;
-  firstName: string;
-  lastName: string;
-}
-
 export const fetchEndUser = async (): Promise<EndUser> => {
-  const result = await request.get('collection-flow/user');
-
-  return result.json<EndUser>();
+  return await request.get('collection-flow/user').json<EndUser>();
 };
-
-export interface CreateEndUserDto {
-  email: string;
-  firstName: string;
-  lastName: string;
-  additionalInfo?: Record<string, unknown>;
-}
 
 export const createEndUserRequest = async ({
   email,
@@ -139,10 +129,27 @@ export const syncContext = async (context: CollectionFlowContext) => {
   return result.json();
 };
 
-export const finalSubmissionRequest = async () => {
+export const updateEndUser = async (data: TUpdateEndUserPluginData) => {
+  const result = await request.post('collection-flow/end-user', {
+    json: data,
+  });
+
+  return result.json();
+};
+
+export const fetchCompanyInformation = async (data: TFetchCompanyInformationPluginData) => {
+  const result = await request.get(`collection-flow/business/business-information`, {
+    searchParams: data,
+  });
+
+  return result.json();
+};
+
+export const finalSubmissionRequest = async (context?: CollectionFlowContext) => {
   const result = await request.post('collection-flow/final-submission', {
     json: {
       eventName: 'COLLECTION_FLOW_FINISHED',
+      context,
     },
   });