+++ apps/kyb-app/src/hooks/useFlowContextQuery/useFlowContextQuery.ts
@@ -1,18 +1,18 @@
-import { useAccessToken } from '@/common/providers/AccessTokenProvider';
 import { collectionFlowQuerykeys } from '@/domains/collection-flow';
 import { useQuery } from '@tanstack/react-query';
 import { HTTPError } from 'ky';
 import { useEndUserQuery } from '../useEndUserQuery';
+import { useWorkflowId } from '@/common/hooks/useWorkflowId';
 
 export const useFlowContextQuery = () => {
-  const { accessToken } = useAccessToken();
+  const workflowId = useWorkflowId();
   const { data: endUser } = useEndUserQuery();
 
   const { data, isLoading, isFetched, error, refetch } = useQuery({
     ...collectionFlowQuerykeys.getContext(endUser?.id ?? null),
     // @ts-ignore
     staleTime: Infinity as const,
-    enabled: !!accessToken,
+    enabled: !!workflowId,
   });
 
   return {