+++ services/workflows-service/scripts/workflows/ui-definition/kyb-with-associated-companies/ui-definition/associated-company-ui-def/associated-ui-definition.ts
@@ -31,7 +31,7 @@ export const definition = {
       {
         name: 'update_end_user',
         pluginKind: 'api',
-        url: `{collectionFlow.config.apiUrl}/api/v1/collection-flow/end-user`,
+        url: `{collectionFlow.config.apiUrl}/api/v1/collection-flow/end-user?workflowId={query.workflowId}`,
         method: 'POST',
         headers: { Authorization: 'Bearer {query.token}' },
         stateNames: [],
@@ -107,7 +107,7 @@ export const definition = {
       {
         name: 'fetch_company_information',
         pluginKind: 'api',
-        url: `{collectionFlow.config.apiUrl}/api/v1/collection-flow/business/business-information`,
+        url: `{collectionFlow.config.apiUrl}/api/v1/collection-flow/business/business-information?workflowId={query.workflowId}`,
         method: 'GET',
         stateNames: [],
         headers: { Authorization: 'Bearer {query.token}' },